requirements.txt
crewai
langchain_google_genai #google gemini models
load_dotenv
crewai_tools